requirements.txt
accelerate==1.10.1
audioread==3.1.0
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
decorator==4.4.2
ExifRead==3.5.1
filelock==3.19.1
Flask==2.3.3
fsspec==2025.10.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.3.2
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.43.0
MarkupSafe==3.0.3
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.2
networkx==3.2.1
nltk==3.8.1
numba==0.60.0
numpy==1.26.4
packaging==26.0
pandas==2.0.3
pillow==11.3.0
platformdirs==4.4.0
pooch==1.9.0
proglog==0.1.12
psutil==7.2.2
pycparser==2.23
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
soundfile==0.13.1
soxr==1.0.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.67.2
transformers==4.57.6
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.5
zipp==3.23.0
